Usando ohook useRoom: abstraindo o código de carregamento de elementos da sala
