+++ src/hooks/useRoom.ts
@@ -1,0 +1,57 @@
+import { useState, useEffect } from "react";
+
+import { database } from '../services/firebase';
+
+type Question = {
+  id: string;
+  author: {
+    name: string,
+    avatar: string
+  },
+  content: string,
+  isAnswered: boolean,
+  isHighlighted: boolean
+}
+
+type FirebaseQuestions = Record<string , {
+  author: {
+    name: string,
+    avatar: string
+  },
+  content: string,
+  isAnswered: boolean,
+  isHighlighted: boolean
+}>
+
+
+export function useRoom(roomId: string) {
+  const [questions, setQuestions] = useState<Question[]>([]);
+  const [title, setTitle] = useState('');
+
+  useEffect(() => {
+    const roomRef = database.ref(`rooms/${roomId}`);
+
+    // escuta apenas uma vez o evento 'value' e então executa a função
+    roomRef.on('value', room => {
+      const databaseRoom = room.val();
+      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
+
+      const parsedQuestions = Object.entries(firebaseQuestions).map(([key, value]) => {
+        return {
+          id: key,
+          content: value.content,
+          author: value.author,
+          isHighlighted: value.isHighlighted,
+          isAnswered: value.isAnswered,
+        }
+      });
+
+      setTitle(databaseRoom.title);
+      setQuestions(parsedQuestions);
+    })
+  }, [roomId]);
+
+  // retornando as informações
+  return { questions, title }; 
+
+}
