+++ src/pages/Room.tsx
@@ -10,32 +10,12 @@ import '../styles/room.scss';
 
 import { useState, FormEvent, useEffect } from 'react';
 import { useAuth } from '../hooks/useAuth';
+import { useRoom } from '../hooks/useRoom';
 
 import { database } from '../services/firebase';
 
-type FirebaseQuestions = Record<string , {
-  author: {
-    name: string,
-    avatar: string
-  },
-  content: string,
-  isAnswered: boolean,
-  isHighlighted: boolean
-}>
-
 type RoomParams = {
   id: string;
-}
-
-type Question = {
-  id: string;
-  author: {
-    name: string,
-    avatar: string
-  },
-  content: string,
-  isAnswered: boolean,
-  isHighlighted: boolean
 }
 
 export function Room() {
@@ -43,33 +23,9 @@ export function Room() {
   const roomId = params.id;
   const [newQuestion, setNewQuestion] = useState('');
   const { user } = useAuth();
-  const [questions, setQuestions] = useState<Question[]>([]);
-  const [title, setTitle] = useState('');
-
-  useEffect(() => {
-    const roomRef = database.ref(`rooms/${roomId}`);
-
-    // escuta apenas uma vez o evento 'value' e então executa a função
-    roomRef.on('value', room => {
-      const databaseRoom = room.val();
-      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
-
-      const parsedQuestions = Object.entries(firebaseQuestions).map(([key, value]) => {
-        return {
-          id: key,
-          content: value.content,
-          author: value.author,
-          isHighlighted: value.isHighlighted,
-          isAnswered: value.isAnswered,
-        }
-      });
-
-      setTitle(databaseRoom.title);
-      setQuestions(parsedQuestions);
-    })
-  }, [roomId]);
-
-  async function handelSendQuestion(event: FormEvent) {
+  const { questions, title } = useRoom(roomId);
+  
+  async function handleSendQuestion(event: FormEvent) {
     event.preventDefault();
 
     // verifica se há algo escrito dentro do newQuestion
@@ -115,7 +71,7 @@ export function Room() {
           )}
         </div>
 
-        <form onSubmit={handelSendQuestion}>
+        <form onSubmit={handleSendQuestion}>
           <textarea 
             placeholder="O que você quer perguntar?"
             value={newQuestion}
